feat: export traces using the otlp protocol and grpc
