+++ Cargo.toml
@@ -14,7 +14,7 @@ dotenv = "0.15.0"
 futures = "0.3.28"
 jsonwebtoken = "8.3.0"
 lapin = { version = "2.2.1", optional = true }
-lettre = "0.10.4"
+lettre = { version = "0.10.4", optional = true }
 libreauth = "0.16.0"
 once_cell = "1.18.0"
 openssl = "0.10.54"
@@ -40,10 +40,10 @@ sqlx = { version = "0.7.1", features = [
 ], optional = true }
 strum = "0.25.0"
 strum_macros = "0.25.0"
-tera = "1.19.0" # template engine
+tera = { version = "1.19.0", optional = true } # template engine
 tokio = { version = "1.28.2", features = ["macros", "rt", "rt-multi-thread"] }
 tonic = { version = "0.9", optional = true } # gRPC
-tracing = { version = "0.1.37" }
+tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.17", optional = true }
 tracing-opentelemetry = { version = "0.20.0", optional = true }
 
@@ -56,22 +56,15 @@ path = "src/lib.rs"
 
 [features]
 default = ["grpc-service", "rest-service"]
-config = []
 grpc = ["prost", "protoc", "tonic"]
-grpc-service = [
-    "grpc",
-    "postgres",
-    "redis-cache",
-    "rabbitmq",
-    "config",
-    "trace",
-]
+grpc-service = ["grpc", "postgres", "redis-cache", "rabbitmq", "smtp", "tracer"]
 postgres = ["sqlx"]
 rabbitmq = ["deadpool-lapin", "lapin"]
 redis-cache = ["redis", "reool"]
 rest = ["actix-web", "redis-cache"]
-rest-service = ["rest", "config", "trace"]
-trace = [
+rest-service = ["rest", "tracer"]
+smtp = ["lettre", "tera"]
+tracer = [
     "tonic",
     "tracing-subscriber",
     "tracing-opentelemetry",
